backend/requirements.txt
asgiref==3.11.0
Django==4.2.27
django-cors-headers==4.9.0
djangorestframework==3.16.1
dnspython==2.8.0
gunicorn==24.1.1
mongoengine==0.29.1
packaging==26.0
pymongo==3.11.4
python-dotenv==1.2.1
pytz==2025.2
sqlparse==0.5.5
tzdata==2025.3
